Update createPickupLZ.sqf
we now take hotlz into consideration before spawning anti-air. anti-air can only spawn if lz is considered hot. this means if player set hot lz to never they cannot end up with a hotlz anyway because the AA chance ended up spawning one.

Now if how LZ is on 50% and ends up hot and AA is on 10% we will only roll the AA if the LZ was hot. This means the odds for AA is effectively halved, might want to increase the odds of AA by 100% in the parameters to compensate.

10% is effectively 5% now, etc.